Replace Image component with useAsset hook
**Why**: Asset mapping may be needed for additional image attributes like `srcset`, `sizes`, etc
